refactor(file): file upload를 위한 exception 추가
