+++ src/main/java/com/github/dankook_univ/meetwork/file/application/minio/MinioService.java
@@ -1,0 +1,13 @@
+package com.github.dankook_univ.meetwork.file.application.minio;
+
+import com.github.dankook_univ.meetwork.file.exceptions.MinioStatObjectException;
+import java.io.InputStream;
+
+public interface MinioService {
+
+    Boolean upload(String key, InputStream inputStream, Long size, String mime);
+
+    Boolean delete(String key);
+
+    Boolean stat(String key) throws MinioStatObjectException;
+}
