+++ src/main/java/com/github/dankook_univ/meetwork/file/exceptions/MinioStatObjectException.java
@@ -1,0 +1,5 @@
+package com.github.dankook_univ.meetwork.file.exceptions;
+
+public class MinioStatObjectException extends Throwable {
+
+}
